move Process to core. major clean up of old functions
