+++ vivarium/processes/degradation.py
@@ -3,7 +3,8 @@ from __future__ import absolute_import, division, print_function
 import os
 import copy
 
-from vivarium.core.process import Process, keys_list
+from vivarium.utils.dict_utils import keys_list
+from vivarium.core.tree import Process
 from vivarium.core.composition import (
     simulate_process,
     plot_simulation_output,
