+++ vivarium/processes/derive_counts.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import, division, print_function
 
 from vivarium.processes.derive_globals import AVOGADRO
-from vivarium.core.process import Deriver
+from vivarium.core.tree import Deriver
 from vivarium.utils.units import units
