+++ vivarium/processes/derive_globals.py
@@ -6,7 +6,7 @@ import math
 from scipy import constants
 import numpy as np
 
-from vivarium.core.process import Deriver
+from vivarium.core.tree import Deriver
 from vivarium.utils.units import units
 from vivarium.utils.dict_utils import deep_merge
