+++ vivarium/processes/homogeneous_environment.py
@@ -4,7 +4,7 @@ import copy
 
 from vivarium.utils.units import units
 from vivarium.utils.dict_utils import deep_merge
-from vivarium.core.process import Deriver
+from vivarium.core.tree import Deriver
 
 from vivarium.processes.derive_globals import AVOGADRO
