+++ vivarium/processes/metabolism.py
@@ -19,7 +19,7 @@ import argparse
 import numpy as np
 import matplotlib.pyplot as plt
 
-from vivarium.core.process import Process
+from vivarium.core.tree import Process
 from vivarium.core.composition import (
     simulate_process_in_experiment,
     save_timeseries,
