+++ vivarium/processes/timeline.py
@@ -3,8 +3,7 @@ from __future__ import absolute_import, division, print_function
 import copy
 
 from vivarium.utils.dict_utils import deep_merge
-from vivarium.core.process import Process
-from vivarium.core.tree import Compartment
+from vivarium.core.tree import Compartment, Process
 
 
 class TimelineProcess(Process):
